LayoutTests/media/media-webm-opus-variable-length-expected.txt


Check that an opus file made of two concatenated segments, one encoded with 40ms packets and the other 20ms can play continously.

EXPECTED (audio.canPlayType('audio/webm; codecs=opus') == 'probably') OK
RUN(audio.src = "content/opus_variable.webm")
EVENT(loadedmetadata)
EXPECTED (audio.duration >= '0.98') OK
RUN(audio.play())
EVENT(ended)
EXPECTED (audio.currentTime >= '0.98') OK
END OF TEST
